requirements.txt
astroid==2.15.6
blinker==1.6.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
identify==2.5.28
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.4.0
PyJWT==2.8.0
pylint==2.17.5
python-dotenv==1.0.0
PyYAML==6.0.1
tomlkit==0.12.1
virtualenv==20.24.5
Werkzeug==2.3.7
wrapt==1.15.0
